Refactored the CERT screen into the individual component pages
