+++ EmergencyDataCollectionApp/package.json
@@ -18,6 +18,7 @@
     ]
   },
   "dependencies": {
+    "@react-navigation/material-top-tabs": "^6.6.5",
     "@testing-library/react-native": "^12.3.2",
     "expo": "~49.0.13",
     "expo-status-bar": "~1.6.0",
